refactor: refactor user profile code
